+++ Integrations/EWSMailSender/EWSMailSender.py
@@ -7,20 +7,30 @@ import warnings
 import traceback
 
 import getpass
-getpass_getuser = getpass.getuser
 
 
 # work arround for bug in exchangelib: https://github.com/ecederstrand/exchangelib/issues/448
-def getuser_no_fail():
-    try:
-        user = getpass_getuser()
-    except KeyError:
-        # getuser() fails on some systems. Provide a sane default.
-        user = 'exchangelib'
-    return user
+class FixGetPass(object):
+    def __init__(self):
+        self.getpass_getuser_org = getpass.getuser
+
+        def getuser_no_fail():
+            # getuser() fails on some systems. Provide a sane default.
+            user = 'ews'
+            try:
+                if self.getpass_getuser_org:
+                    user = self.getpass_getuser_org()
+            except KeyError:
+                pass
+            return user
+        getpass.getuser = getuser_no_fail
+
+    def __del__(self):
+        if self.getpass_getuser_org and getpass:
+            getpass.getuser = self.getpass_getuser_org
 
 
-getpass.getuser = getuser_no_fail
+_fix_getpass = FixGetPass()
 
 warnings.filterwarnings("ignore")
 
@@ -30,6 +40,7 @@ log_handler = None
 
 
 def start_logging():
+    logging.raiseExceptions = False
     global log_stream
     global log_handler
     if log_stream is None:
@@ -51,7 +62,7 @@ IS_TEST_MODULE = False
 
 # load arguments
 USE_PROXY = demisto.params().get('proxy', False)
-NON_SECURE = demisto.params().get('insecure', True)
+NON_SECURE = demisto.params().get('insecure', False)
 AUTH_METHOD_STR = demisto.params().get('authType', 'Basic').lower()
 EWS_SERVER = demisto.params().get('ewsServer', 'https://outlook.office365.com/EWS/Exchange.asmx/')
 VERSION_STR = demisto.params().get('defaultServerVersion', '2013')
@@ -248,7 +259,7 @@ config = None  # type: ignore
 
 
 def main():
-    global USERNAME, PASSWORD, ACCOUNT_EMAIL
+    global USERNAME, PASSWORD, ACCOUNT_EMAIL, log_stream
     USERNAME = demisto.params()['credentials']['identifier']
     PASSWORD = demisto.params()['credentials']['password']
     ACCOUNT_EMAIL = demisto.params().get('mailbox', None)
@@ -308,6 +319,7 @@ def main():
             try:
                 logging.getLogger().removeHandler(log_handler)  # type: ignore
                 log_stream.close()
+                log_stream = None
             except Exception as ex:
                 demisto.error("EWS Mail Sender: unexpected exception when trying to remove log handler: {}".format(ex))