+++ Integrations/ThreatConnect/CHANGELOG.md
@@ -1,5 +1,17 @@
 ## [Unreleased]
 Added 8 new commands.
+   - ***tc-get-group***
+   - ***tc-get-group-attributes***
+   - ***tc-get-group-security-labels***
+   - ***tc-get-group-tags***
+   - ***tc-download-document***
+   - ***tc-get-group-indicators***
+   - ***tc-get-associated-groups***
+   - ***tc-associate-group-to-group***
+
+
+## [19.8.2] - 2019-08-22
+Added 8 new commands.
   - ***tc-get-groups***
   - ***tc-add-group-security-label***
   - ***tc-add-group-tag***